pyproject.toml
[project]
name = "mcp-server-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name="Amit Singh", email = "write.amitsingh@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.9.4",
    "mysql-connector-python>=9.2.0",
]

[project.scripts]
mcp-server-mysql = "mcp_server_mysql:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
